+++ client/src/index.js
@@ -6,12 +6,15 @@ import App from './App';
 import { AlertProvider } from './contexts/AlertContext';
 import { UserProvider } from './contexts/UserContext';
 import { ProfileProvider } from './contexts/ProfileContext';
+import { ListsProvider } from './contexts/ListsContext';
 
 ReactDOM.render(
     <AlertProvider>
         <UserProvider>
             <ProfileProvider>
-                <App />
+                <ListsProvider>
+                    <App />
+                </ListsProvider>
             </ProfileProvider>
         </UserProvider>
     </AlertProvider>,